Added descriptor count to shader bindings
