+++ include/shaders.hpp
@@ -1,0 +1,18 @@
+#pragma once
+
+
+#include "ShaderBuilder.hpp"
+
+using GlobalShaderTypes = ShaderTypes<
+  Shader<
+    "main_vert",
+    "bin/vert.spv",
+    SHADER_VERTEX
+  >,
+  Shader<
+    "main_frag",
+    "bin/frag.spv",
+    SHADER_FRAGMENT,
+    ImageBinding<1, 1>
+  >
+>;
